ADD: Invincibility
When player enters a room, they have invincibility until the start of the second turn. This means any enemy trying to capture the player in the enemy's first turn is destroyed. Note this is NOT a 'capture' so the player doesn't get the piece added to their chain. - DONE

When the player is captured, we do NOT reload the scene. This eats up too much time and is too boring. Instead, the player loses a life and destroys the enemy who tried to capture them. However if the player is on 0 lives they can no longer use this ability. If the player is captured while on 0 lives then we show the game over. - DONE